feat: Introduce dashboard caching utility functions for improved data management
